Fixed bug when segment has several surfaces and other changes.
- Fix writing bug when surface count > 1
- Compute Surface Count if it is undefined
- Read and write surface processing attributes
- Fix minor bugs

+++ Source/MediaStorageAndFileFormat/gdcmSegmentWriter.cxx
@@ -301,7 +301,7 @@ bool SegmentWriter::PrepareWrite()
     // else assert? return false? gdcmWarning?
 
     //*****   Surface segmentation    *****//
-    const unsigned int surfaceCount = segment->GetSurfaceCount();
+    const unsigned long surfaceCount = segment->GetSurfaceCount();
     if (surfaceCount > 0)
     {
       // Surface Count
